Split out DictQueryset into its own file
