+++ brubeck/queryset/__init__.py
@@ -1,3 +1,4 @@
-from brubeck.queryset.base import AbstractQueryset, DictQueryset
+from brubeck.queryset.base import AbstractQueryset
+from brubeck.queryset.dict import DictQueryset
 from brubeck.queryset.redis import RedisQueryset
